feat: add workflows
feat: remove typer

feat: cleanup vitepress theme

feat: extract sidebar

feat: after-install scripts

feat: pass wd

feat: add shortcuts

fix: deadlinks

feat: add missing articles

feat: keep source when symlinking

feat: revert package.json

feat: cleanup

vercel

fix: build

fix: logo

feat: algolia

feat: remove typer
