+++ .github/scripts/docs.yml
@@ -1,0 +1,1 @@
+build-end: {}
