+++ apps/docs/package.json
@@ -3,6 +3,7 @@
   "private": true,
   "version": "1.7.1",
   "description": "",
+  "type": "module",
   "scripts": {
     "dev": "vitepress",
     "build": "vitepress build",
@@ -16,13 +17,14 @@
     "markdown-it": "14.1.0",
     "vitepress": "1.0.0-beta.7",
     "vitepress-plugin-search": "1.0.4-alpha.22",
-    "vitepress-shopware-docs": "0.3.0-beta.44",
+    "vitepress-shopware-docs": "^1.1.41",
     "vue": "3.4.30"
   },
   "devDependencies": {
     "@types/markdown-it": "14.1.1",
     "@types/node": "20.14.8",
     "@vitejs/plugin-vue": "5.0.5",
+    "sass": "^1.77.6",
     "find-in-files": "0.5.0",
     "ts-dox": "0.1.0",
     "typescript": "5.5.2",
