+++ apps/docs/src/getting-started/templates/astro-template.md
@@ -23,7 +23,7 @@ npm i && npm run dev
 
 ## Configure
 
-The blank template is pre-configured to connect to a public Shopware backend, so you can follow our [building guides](/getting-started/) right away.
+The blank template is pre-configured to connect to a public Shopware backend, so you can follow our [building guides](../../getting-started/) right away.
 
 In order to connect it to your own store backend, you need to edit the `env.development`:
