+++ apps/docs/src/packages/api-client/ClientSettings.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# ClientSettings
+
+<!-- PLACEHOLDER_DESCRIPTION -->
