+++ apps/docs/src/packages/api-client/ConfigChangedArgs.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# ConfigChangedArgs
+
+<!-- PLACEHOLDER_DESCRIPTION -->
