+++ apps/docs/src/packages/api-client/CustomerRegisterResponse.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# CustomerRegisterResponse
+
+<!-- PLACEHOLDER_DESCRIPTION -->
