+++ apps/docs/src/packages/api-client/CustomerResetPasswordParam.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# CustomerResetPasswordParam
+
+<!-- PLACEHOLDER_DESCRIPTION -->
