+++ apps/docs/src/packages/api-client/CustomerUpdateEmailParam.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# CustomerUpdateEmailParam
+
+<!-- PLACEHOLDER_DESCRIPTION -->
