+++ apps/docs/src/packages/api-client/CustomerUpdatePasswordParam.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# CustomerUpdatePasswordParam
+
+<!-- PLACEHOLDER_DESCRIPTION -->
