+++ apps/docs/src/packages/api-client/CustomerUpdateProfileParam.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# CustomerUpdateProfileParam
+
+<!-- PLACEHOLDER_DESCRIPTION -->
