+++ apps/docs/src/packages/api-client/GetStoreNavigationParams.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# GetStoreNavigationParams
+
+<!-- PLACEHOLDER_DESCRIPTION -->
