+++ apps/docs/src/packages/api-client/NewsletterSubscribeData.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# NewsletterSubscribeData
+
+<!-- PLACEHOLDER_DESCRIPTION -->
