+++ apps/docs/src/packages/api-client/ShopwareApiInstance.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# ShopwareApiInstance
+
+<!-- PLACEHOLDER_DESCRIPTION -->
