+++ apps/docs/src/packages/api-client/getProductCrossSelling.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# getProductCrossSelling
+
+<!-- PLACEHOLDER_DESCRIPTION -->
