+++ apps/docs/src/packages/api-client/getSitemap.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# getSitemap
+
+<!-- PLACEHOLDER_DESCRIPTION -->
