+++ apps/docs/src/packages/api-client/getSitemapEndpoint.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# getCmsPage
+
+<!-- PLACEHOLDER_DESCRIPTION -->
