+++ apps/docs/src/packages/api-client/setCurrentCountry.md
@@ -1,0 +1,7 @@
+---
+category: api-client
+---
+
+# setCurrentCountry
+
+<!-- PLACEHOLDER_DESCRIPTION -->
