+++ apps/docs/src/packages/helpers/UiMediaGalleryItem.md
@@ -1,0 +1,7 @@
+---
+category: helpers
+---
+
+# UiMediaGalleryItem
+
+<!-- PLACEHOLDER_DESCRIPTION -->
