+++ apps/docs/src/packages/helpers/UiProductProperty.md
@@ -1,0 +1,7 @@
+---
+category: helpers
+---
+
+# UiProductProperty
+
+<!-- PLACEHOLDER_DESCRIPTION -->
