+++ package.json
@@ -27,7 +27,10 @@
     "taze": "taze major -wr",
     "prepare": "husky",
     "generate-types": "turbo run generate-types",
-    "generateDependencyChangelog": "esno ./scripts/generateDependencyChangelog.ts"
+    "generateDependencyChangelog": "esno ./scripts/generateDependencyChangelog.ts",
+    "docs:env": "[ -d \"../developer-portal\" ] && ../developer-portal/docs-cli.cjs pull || (git clone git@github.com:shopware/developer-portal.git ../developer-portal && pnpm i -C ../developer-portal)",
+    "docs:link": "../developer-portal/docs-cli.cjs link --src apps/docs/src --dst frontends --symlink --keep",
+    "docs:preview": "../developer-portal/docs-cli.cjs preview"
   },
   "devDependencies": {
     "@biomejs/biome": "1.8.1",
